+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## [Unreleased]
 
+### Changed
+
+- [BREAKING] Consolidated `spec.clusterConfig.authenticationConfig` to `spec.clusterConfig.authentication` which now takes a vector of AuthenticationClass references  ([#303]).
+
+[#303]: https://github.com/stackabletech/airflow-operator/pull/303
+
 ## [23.7.0] - 2023-07-14
 
 ### Added